pyproject.toml
[tool.poetry]
name = "screenplay_pdf_to_json"
version = "0.1.5"
description = ""
authors = ["VVNoodle <brickkace@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
"pdfminer.six" = "^20200124"
spacy = "^2.2.4"

[tool.poetry.dev-dependencies]
pytest = "^5.4.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
